python-server/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.10.0
pymongo>=4.6.1
python-dotenv>=1.0.0
openai>=1.10.0
langchain>=0.1.0
langchain-google-genai>=0.0.6
pinecone-client>=3.0.0
requests>=2.31.0
httpx>=0.26.0
aiohttp>=3.9.0
azure-storage-blob>=12.19.0
python-multipart>=0.0.9
